+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Added
+- return of from_topas3 method from older versions of attune, for backcompat
+
 ## [0.5.0]
 
 ### Added